+++ internal/configuration/set_test.go
